added test config, removed serde dependency, fixed bugs found while writing tests

+++ src/lib.rs
@@ -58,7 +58,7 @@ macro_rules! enum_with_custom {(
         $(#[$meta])*
         $vis enum $name {
             $($member),*,
-            Custom(usize)
+            Custom(u32)
         }
 
         impl $name {
